api: Fill the CRD definition with ovn-kubernetes CNI relevant fields
The following CNI fields are *not* exposed:
- `cniVersion`: Reflects the CNI spec version,
  added to the NAD config when created.
- `type`: The ovn-kubernetes CNI plugin binary,
  added to the NAD config when created.
-  `topology`: Set to `layer2` in the NAD config
   to reflect layered network.
-  `vlanID`: The ovn-kubernetes CNI parameter which is not used
   with the overlay network topology.

The logical network name inputted through the API (`name` field) is
planned to be prefixed with the CR namespace when generating the
backend NetworkAttachmentDefinition. This aims to reduce the chances
of different users from different namespaces to have access to the same
logical network.
However, this is still not safe, as it may collide with NAD
configuration that the cluster admin may have created.

Signed-off-by: Edward Haas <edwardh@redhat.com>