controller, reconciler: Create NetworkAttachmentDefinition
On update/create of OverlayNetwork, create corresponding
NetworkAttachmentDefinition.

The NetworkAttachmentDefinition is created with OwnerReference [1] pointing to
the reconciled object.
Once the reconciled object is deleted the corresponding net-attach-def object
is deleted as well by the cluster garbage collector.

This change does not set the net-attach-def spec.config, it will be implemented
in follow up commit.

Do not fail when the reconciled OverlayNetwork object is it not found as it may
got deleted just before getting into the reconcile loop.

[1] https://sdk.operatorframework.io/docs/building-operators/ansible/reference/retroactively-owned-resources/

Signed-off-by: Or Mergi <ormergi@redhat.com>